This PR expands the API of UnsignedBigInteger to include :

- Static non-allocating add/subtract/multiply/divide/shift operations
- Special cases for operations between a BigInt and a small number (currently implemented : `divide_by_u16`)

These improvements are then used in ModularFunction to improve performances.

The method signatures are very unwieldy, but contain everything required to prevent allocations. Names can be reworked/changed as needed !

This PR also includes a bookkeeping commit (added a missing `move()`, fixed a typo in the tests names, and moved some methods around in the UnsignedBigInteger to make them thematically closer).

Here are the performance improvements (computed very professionally by sitting next to the computer with a timer and hitting it when the test ends) :

| Test | Original code | Using `without_alloc` variants |
|----------------------------|----------------|-----------------------|
| RSA PKCS #1 1.5 Encryption | 6mn 15s | 4mn 03s (-35%) |
| RSA Encrypt | 5mn 22s | 4mn 24s (-12%) |

Here are the top 4 methods with self-hits (inverted tree) during a 5-second `profile` run in the middle of the `RSA PKCS #1 1.5 Encryption` test :

| Method | Original code | Using `without_alloc` variants |
|----------------------------|----------------|-----------------------|
| `multiply_by{_without_allocation}` | 1182 | 1702 |
| `divide_by{_without_allocation}` | 1289 | 1498 |
| `malloc` | 1086 | 31 |
| `free` | 1085 | 22 |