@Dexesttp 2020-05-02
I've started to work on more efficient algorithms, and have renamed the PR to express that. It's also set as draft until these optimizations are ended, though it could probably be merged whenever.

For now, I started with u16 multiplication to allow for simplified computations. It should slightly improve the execution time for RSA decryption (by changing the `ep` shift from O(len(ep)^2) to O(len(ep))), but I haven't re-timed it to check the actual improvement.

To see where the remaining bottlenecks are during `pk` computation, I also started to do some specific testing. It showed an even split between the multiplication and division steps, which makes sense since both algorithms are O(N^2).  So, the next steps will be to improve the general cases for multiply and divide computations.

I'll start by implementing Karatsuba multiplication and see if that improves the computation speed.